requirements.txt


fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.26.0
motor==3.3.2
pymongo==4.6.1
